delete direcory with notes
Видалив папку з нотатками а її вміст перекинув до загального проєкту
