+++ notes.py
